+++ Changelog.md
@@ -16,6 +16,7 @@
 - Change: The Fake Lava/Water Buckets are now created in a brewing stand, by putting the potion in the top slot, and the bucket(s) in the bottom slot(s)
 - Change: Sentries can now be placed in water
 - Change: Massively increased explosion resistance of all blocks to hopefully prevent them from being broken by modded explosions
+- Change: Placing a Laser Block and Inventory Scanner now also places the fields if replaceable blocks (like grass or snow) are in the way
 - API: Split IPasswordProtected into two interfaces (IPasswordProtected and ICodebreakable), meaning blocks can now be hacked by the codebreaker without needing to be password protected
 - API: Removed IPasswordProtected#isCodebreakable
 - API: New linked action: OWNER_CHANGED